requirements.txt
langchain==0.0.335
nltk==3.8.1
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.28.2
streamlit_chat==0.1.1
transformers==4.35.0
